misc changes

+++ platformticketdispensingsystem/src/main/java/com/ptds/GuestDetails.java
@@ -42,7 +42,7 @@ class GuestDetails {
     JFormattedTextField PhoneNumber;
     static String PNR;
 
-    static int amount = 20;
+    static float amount = 20;
 
     GuestDetails(JTextField name , JFormattedTextField PhNo){
